\name{print.repsample_search_result}
\alias{print.repsample_search_result}
\title{Print a \code{repsample_search_result}}
\usage{
\method{print}{repsample_search_result}(x, ...)
}
\arguments{
\item{x}{A \code{repsample_search_result} object.}

\item{...}{Unused.}
}
\description{
Print a \code{repsample_search_result}
}
